ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
deepface==0.0.75
opencv-python==4.8.1.78
tensorflow==2.14.0
numpy==1.24.3
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
openai==1.3.0
transformers==4.35.0
torch==2.1.1
